feat: refactor the app to isolate files
