+++ src/modules/accommodation/controllers/rooms-controller.js
@@ -1,7 +1,6 @@
 const fs = require('fs');
 const { validate } = require('../../../utils/validate');
 const { Room, createRoomSchema, Accommodation } = require('../models');
-const sequelize = require('../../../config/SequelizeConfig');
 const assert = require('http-assert');
 
 /**
@@ -99,57 +98,6 @@ class RoomsController {
         }
       });
     }
-  }
-
-  /**
-   * @param {Express.Request} req
-   * @param {Express.Response} res
-   * @returns {*} cooresponding rooms
-   */
-  static async uploadImage(req, res) {
-    const room = await Room.findByPk(req.params.id);
-
-    if (!room) {
-      return res.status(404).json({
-        status: 'NOT_FOUND',
-        errors: {
-          request: [
-            'Room with this id is not found'
-          ]
-        }
-      });
-    }
-
-    const { files } = req;
-    if (!files.length) {
-      return res.status(404).json({
-        status: 'BAD_REQUEST',
-        errors: {
-          files: [
-            'No Files Uploaded'
-          ]
-        }
-      });
-    }
-
-    const uploader = (path) => cloudinary.uploads(path);
-
-    const urls = [];
-
-    await Promise.all(files.map(async (file) => {
-      const { path } = file;
-      const newPath = await uploader(path);
-      urls.push(newPath);
-      fs.unlinkSync(path);
-    }));
-
-    await Room.update({ image_paths: urls.map((url) => url.url).join(',') }, { where: { id: room.id } });
-
-    return res.status(200).json({
-      status: 'SUCCESS',
-      room,
-      files
-    });
   }
 }
