+++ src/modules/accommodation/models/accommodation.js
