+++ src/modules/accommodation/routes/accommodation-routes.js
@@ -1,7 +1,6 @@
 const express = require('express');
 const { authenticate } = require('../../../middlewares/authenticate');
 const { authorize } = require('../../../middlewares/authorize');
-const upload = require('../../../utils/multer');
 
 const router = express.Router();
 
@@ -63,31 +62,5 @@ router.get('/:id', AccomodationsController.getById);
  *        description: Unauthorized to create an accommodation
  */
 router.post('/', [authenticate, authorize('ADMIN')], AccomodationsController.create);
-
-/**
- * @swagger
- * /api/v1/accommodations/{id}/upload-image:
- *  put:
- *    tags:
- *      - Accommodation
- *    description: Upload images of an accomodation
- *    parameters:
- *      - name: files
- *        in: formData
- *        type: file
- *        description: The file to upload
- *      - name: id
- *        in: params
- *        type: number
- *        description: The id of the accomodation
- *    responses:
- *      200:
- *        description: Images uploaded
- *      400:
- *        description: Bad Request
- *      401:
- *        description: Unauthorized to create an accommodation
- */
-router.put('/:id/upload-image', [authenticate, authorize('ADMIN'), upload.array('files')], AccomodationsController.uploadImage);
 
 module.exports = router;
