+++ src/modules/accommodation/routes/index.js
