+++ src/modules/file/controllers/index.js
@@ -1,0 +1,5 @@
+const { FileController } = require("./file-controller")
+
+module.exports = {
+  FileController
+}
