+++ src/modules/file/routes/index.js
@@ -1,0 +1,6 @@
+const fileRouter = require("./file-routes")
+
+
+module.exports = {
+  fileRouter
+}
